refactor: 워크플로우 목록 페이지 구조 분리
