+++ src/pages/workflows/model/constants.ts
@@ -1,0 +1,7 @@
+export const WORKFLOW_FILTERS = [
+  { key: "all", label: "전체" },
+  { key: "active", label: "실행" },
+  { key: "inactive", label: "중지됨" },
+] as const;
+
+export const WORKFLOW_LIST_PAGE_SIZE = 20;
